refactor: use pinia webhook store instead of vuex store
